+++ example/lib/home_page.dart
@@ -1,4 +1,5 @@
 import 'package:example/card_page.dart';
+import 'package:example/text_fields_page.dart';
 import 'package:example/text_page.dart';
 import 'package:flutter/material.dart';
 import 'package:ui_utility_package/enums.dart';
@@ -74,6 +75,34 @@ class _HomePageState extends State<HomePage> {
                   uiUtilityPackage.customText(
                     text:
                         'Double tap to go to a page where examples related to different variations of the custom card widget can be found.',
+                    fontSize: TextSize.subTitle,
+                    color: TextColors.secondary,
+                  ),
+                ],
+              ),
+            ),
+          ),
+
+          ///Text Field widgets
+          GestureDetector(
+            onDoubleTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(builder: (context) => FormFieldsPage()),
+              );
+            },
+            child: uiUtilityPackage.customCard(
+              color: Color(0xff616060),
+              widget: Column(
+                children: [
+                  uiUtilityPackage.customText(
+                    text: 'Text Field Examples',
+                    fontSize: TextSize.title,
+                    color: TextColors.secondary,
+                  ),
+                  uiUtilityPackage.customText(
+                    text:
+                        'Double tap to go to a page where examples related to different variations of the custom form field widget can be found.',
                     fontSize: TextSize.subTitle,
                     color: TextColors.secondary,
                   ),